==> R/nodscov2/7.scenarios-analysis.Rmd <==
---
title: 'scenarios-analysis'
author: 'Olivier GAUFRÈS'
date: '2024-07-30'
output: html_document
---

```{r}
#### LIBRARIES
## DATA MANAGEMENT
library(dplyr)
library(tidyr)
library(purrr)
library(data.table)
library(lubridate)
```


```{r}
##PATHS
wd <- getwd()
out_path <- file.path(wd,'..','..', 'out')
loc_path <- file.path(out_path,'loc-nodscov2')
scenarios_path <- file.path(out_path, 'scenarios-analysis')

## LOAD FUNCTIONS
source(file.path(wd, 'scenarios-analysis-supp.R'))
```

## LOAD LOCALIZATION DATA & GLOBAL PARAMETERS
```{r}
load(file.path(loc_path, 'dev-localization-nodscov2.RData'))
load(file.path(scenarios_path, "parameters-admission-nodscov2.RData"))

## CATEGORY GROUPING INTO TYPE
cat_paramedical <- c("nurse", "student nurse", "aux nurse")
cat_medical <- c("physician", "ext physician", "reeducation staff", "other")

## IDS BY CATEGORY
id_patient <- admission_sim %>% filter(info == 0) %>% distinct(id) %>% pull()
id_hcw <- admission_sim %>% filter(info == 1) %>% distinct(id) %>% pull()
## FOR MEDICAL/PARAMEDICAL WE USE ADMISSION (INDIVIDUALS BEFORE ADDING NEW PATIENTS)
id_paramedical <- admission %>% filter(cat %in% cat_paramedical) %>% distinct(id) %>% pull()
id_medical <- admission %>% filter(cat %in% cat_medical) %>% distinct(id) %>% pull()
```


## BETA/NU COUPLES -> READ SIMULATION RESULTS
```{r}
## ALL COUPLES INSIDE A LIST, EVERY SIMULATION INSIDE SUB-LIST NAMED sim<beta>_<nu> --> WILL BE NAMED A COUPLE
list_dirs <- list.dirs(path = file.path(scenarios_path, 'scenarios-simulations'), recursive = F, full.names = F)
list_sim <- list()

## READ ALL SIMULATION FOR ALL AVAILABLE COUPLES
for (dir in list_dirs){
  list_sim[[dir]] <- list()
  sim_path <- file.path(scenarios_path, 'scenarios-simulations', dir, 'results')
  rdata_files <- list.files(sim_path, pattern = "^sim_.*\\.RData$", full.names = F)
  invisible(lapply(rdata_files, function(file) load_rdata_to_list(file, list_sim, dir)))
}
```

## FOR EACH SIMULATION -> COMPUTE SAR
```{r}
## GET SARs FOR ALL COUPLES
list_SAR <- get_all_SAR(list_sim = list_sim, 
                        id_patient = id_patient, 
                        id_hcw = id_hcw, 
                        id_paramedical = id_paramedical, 
                        id_medical = id_medical)
```

## SAR METRICS
```{r}
# all_SAR_g_metrics <- get_all_SAR_g_metrics(list_SAR = list_SAR) ## If only interested in global SAR
all_SAR_metrics <- compute_all_SAR_metrics(list_SAR = list_SAR)
```

## EPIDEMIC DURATION
```{r}
list_epidemic_duration <- get_all_epidemic_duration(list_sim = list_sim)
all_epidemic_duration_metrics <- get_all_epidemic_duration_metrics(list_epidemic_duration = list_epidemic_duration)
```

## MAX N INFECTIOUS AT PEAK
```{r}

```

## ALL METRICS IN ONE DF
```{r}
all_metrics <- left_join(all_SAR_metrics, all_epidemic_duration_metrics, by="couple", suffix=c("",".y")) %>%
  select(-ends_with(".y"))

write.csv2(all_metrics, file = file.path(scenarios_path, 'scenarios-all_metrics.csv'), row.names = F)
```
